backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# AWS SDK
boto3>=1.29.0
botocore>=1.32.0

# Optional: for enhanced validation
python-multipart>=0.0.6

sqlalchemy>=2.0
pymysql>=1.1
python-dotenv>=1.0

# OAuth and Authentication
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
itsdangerous>=2.1.2


